Update chức năng tìm kiếm, hiện user, hiện trạng thái

+++ Project/src/main/java/com/example/demo/service/CategoryServiceImp.java
@@ -1,6 +1,7 @@
 package com.example.demo.service;
 
 import com.example.demo.dao.CategoryDAO;
+import com.example.demo.dao.ProductDAO;
 import com.example.demo.entity.Category;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,10 +12,12 @@ import java.util.List;
 @Service
 @Transactional
 public class CategoryServiceImp implements CategoryService{
+    private ProductDAO productDAO;
     private CategoryDAO categoryDAO;
     @Autowired
-    public CategoryServiceImp(CategoryDAO categoryDAO) {
+    public CategoryServiceImp(CategoryDAO categoryDAO,ProductDAO productDAO) {
         this.categoryDAO = categoryDAO;
+        this.productDAO = productDAO;
     }
 
     @Override
@@ -34,6 +37,14 @@ public class CategoryServiceImp implements CategoryService{
 
     @Override
     public void deleteById(int id) {
+        long productCount = productDAO.countByCategoryId(id);
+
+        if (productCount > 0) {
+            throw new RuntimeException(
+                    "Danh mục đang có " + productCount + " sản phẩm, không thể xóa"
+            );
+        }
+
         categoryDAO.deleteById(id);
     }
     public long countAll() {
